+++ MinimapButton.lua
@@ -59,8 +59,3 @@ end
 
 local function OnDragStop(self)
     self:StopMovingOrSizing()
-    self.isMoving = false
-
-    -- Calculate angle from minimap center
-    local mx, my = Minimap:GetCenter()
-    local bx, by = self:GetCenter()